+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## New version
 
 * We don't need to add `config.action_mailer.default_url_options` anymore
+* Make gem add bcrypt more resilient
 
 ## Authentication Zero 4.0.2 ##